feat: Update color variables and font size
This commit updates the color variables and font size in the CSS files. It changes the `--thm-white-rgb` value in `color-1.css` to `255, 255, 255` and the `--thm-gray-bg-rgb` value in `color-2.css` to `247, 247, 247`. Additionally, it reduces the font size in the `header.css` file from `14px` to `12px`.

The commit also comments out several sections in the `page.js` file, including `ProjectSection`, `Team`, `BlogSection`, and `BrandSection`. These sections are temporarily disabled.

Finally, it updates the styling of the `BrandLogo` component by adding inline styles to the `h2` element. The color is set to `#ffffff`, the font weight is set to `bold`, the font size is set to `24px`, the text is transformed to uppercase, and there is additional letter spacing. The margin top and bottom are also adjusted.

This commit improves the color scheme and font size consistency in the project, as well as temporarily disables certain sections on the About page. It also enhances the styling of the BrandLogo component.