feat: Update color variables, font size, and component styling
This commit updates the color variables and font size in the CSS files to improve the color scheme consistency. It changes the `--thm-white-rgb` value to `255, 255, 255` and the `--thm-gray-bg-rgb` value to `247, 247, 247`. Additionally, it reduces the font size in the `header.css` file from `14px` to `12px`.

The commit also temporarily disables several sections in the `page.js` file, including `ProjectSection`, `Team`, `BlogSection`, and `BrandSection`.

Furthermore, it enhances the styling of the `BrandLogo` component by adding inline styles to the `h2` element. The color is set to `#ffffff`, the font weight is set to `bold`, the font size is set to `24px`, the text is transformed to uppercase, and there is additional letter spacing. The margin top and bottom are also adjusted.

These changes improve the color scheme and font size consistency in the project, as well as temporarily disable certain sections on the About page. Additionally, it enhances the styling of the BrandLogo component.